pyproject.toml
[project]
name = "rice"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ansible>=11.9.0",
    "whiptail-dialogs>=0.4.1",
]
